+++ server/auth/limits.ts
@@ -1,34 +1,12 @@
 import { check } from "diskusage";
 import { platform } from "os";
+import { simplifiedByteCount } from "../transformers/index.js";
 
 const rootPath = platform() === "win32" ? "c:" : "/";
 
 async function availableSpace(): Promise<number> {
 	const { available } = await check(rootPath);
 	return available;
-}
-
-function fileSizeStringFromByteCount(num: number): string {
-	if (typeof num !== "number" || Number.isNaN(num)) {
-		throw new TypeError("Expected a number");
-	}
-
-	const neg = num < 0;
-	const units = ["B", "kB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"];
-
-	if (neg) {
-		num = -num;
-	}
-
-	if (num < 1) {
-		return `${neg ? "-" : ""}${num} B`;
-	}
-
-	const exponent: number = Math.min(Math.floor(Math.log(num) / Math.log(1000)), units.length - 1);
-	const unit: string = units[exponent] ?? "";
-	const size: string = (num / 1000 ** exponent).toFixed(2);
-
-	return `${neg ? "-" : ""}${size} ${unit}`;
 }
 
 const defaultMaxUsers = 5;
@@ -39,9 +17,7 @@ const totalSpace = await availableSpace();
 export const spacePerUser = totalSpace / MAX_USERS;
 
 console.log(
-	`System has ${fileSizeStringFromByteCount(
-		totalSpace
-	)} available. That's ${fileSizeStringFromByteCount(
+	`System has ${simplifiedByteCount(totalSpace)} available. That's ${simplifiedByteCount(
 		spacePerUser
 	)} for each of our ${MAX_USERS} max users.`
 );
