+++ src/transformers/compactMap.ts
